[#47] Make sure annotations are included in the checks if the annotated element is, too.

+++ revapi-java/src/test/resources/v2/annotations/Attributes.java
@@ -27,5 +27,5 @@ public abstract class Attributes {
 
     @Anno(a = 1, b = 2)
     @Deprecated
-    public abstract void m(int i);
+    public abstract void m(float i);
 }
